`Truncate` moved to top-level component
(no functional change, just prevents build warning)
