+++ packages/components/src/Text/index.ts
@@ -31,4 +31,3 @@ export * from './Paragraph'
 export * from './ReplaceText'
 export * from './Span'
 export * from './Text'
-export * from './Truncate'
