+++ packages/components/src/Tree/TreeGroup.tsx
@@ -28,7 +28,7 @@ import React, { FC } from 'react'
 import styled, { css } from 'styled-components'
 import { color, TextColorProps } from '@looker/design-tokens'
 import { AccordionDisclosure } from '../Accordion'
-import { Truncate } from '../Text'
+import { Truncate } from '../Truncate'
 import { TreeItemLabel } from './TreeItem'
 
 export interface TreeGroupProps extends TextColorProps {
