+++ packages/components/src/Truncate/useIsTextTruncated.ts
